requirements.txt
anyio==4.9.0
blinker==1.9.0
certifi==2025.6.15
click==8.2.1
Flask==3.1.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pony==0.7.19
python-nmap==0.7.1
sniffio==1.3.1
typing_extensions==4.14.0
Werkzeug==3.1.3
